@@ -1,30 +1,30 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>net8.0</TargetFramework>
+        <TargetFramework>net9.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
     </PropertyGroup>
 
     <ItemGroup>
       <PackageReference Include="LargeXlsx" Version="1.11.1" />
+      <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="9.0.1" />
+      <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="9.0.1" />
     </ItemGroup>
 
     <ItemGroup>
-      <None Update="TemplateBuilding\Lonira\template.json">
+      <None Remove="TemplateBuilding\Suliver\template.json" />
+      <EmbeddedResource Include="TemplateBuilding\Suliver\template.json">
         <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-      </None>
-      <None Update="TemplateBuilding\Suliver\template.json">
+      </EmbeddedResource>
+      <None Remove="TemplateBuilding\MegaTrading\template.json" />
+      <EmbeddedResource Include="TemplateBuilding\MegaTrading\template.json">
         <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-      </None>
-      <None Update="TemplateBuilding\MegaTrading\template.json">
+      </EmbeddedResource>
+      <None Remove="TemplateBuilding\Lonira\template.json" />
+      <EmbeddedResource Include="TemplateBuilding\Lonira\template.json">
         <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-      </None>
-    </ItemGroup>
-
-    <ItemGroup>
-      <Compile Remove="DataAccess\CosmosDbConfiguration.cs" />
-      <Compile Remove="DataAccess\CosmosDbContext.cs" />
+      </EmbeddedResource>
     </ItemGroup>
 
 </Project>
