@@ -3,7 +3,7 @@ using System.Runtime.CompilerServices;
 
 namespace Kroiko.Domain.CellsExtracting;
 
-public record Detail(
+public sealed record Detail(
     double Height,
     double Width,
     int Quantity,
@@ -29,17 +29,4 @@ public record Detail(
     double OversizingWidth) : INotifyPropertyChanged
 {
     public event PropertyChangedEventHandler? PropertyChanged;
-
-    protected virtual void OnPropertyChanged([CallerMemberName] string? propertyName = null)
-    {
-        PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
-    }
-
-    protected bool SetField<T>(ref T field, T value, [CallerMemberName] string? propertyName = null)
-    {
-        if (EqualityComparer<T>.Default.Equals(field, value)) return false;
-        field = value;
-        OnPropertyChanged(propertyName);
-        return true;
-    }
 }
