@@ -2,5 +2,5 @@
 
 public interface ITemplateBuilder
 {
-    public Task<IList<ISheet>> BuildTemplateAsync(ContactInfo contactInfo, IEnumerable<KroikoFile> files);
+    public Task<IList<ISheet>> BuildTemplateAsync(IEnumerable<KroikoFile> files);
 }
