@@ -3,5 +3,5 @@
 public class MegaTradingFileNameProvider: IFileNameProvider
 {
     public string GetFileNameForSheet(ISheet sheet) =>
-        $"{DateTime.Now:yyyy-MM-dd}_{{CompanyName}}.xlsx";
+        $"{DateTime.Now:yyyy-MM-dd}_{TemplateBuilderBase.CompanyNameCellFlag}.xlsx";
 }
