@@ -4,6 +4,6 @@ public class SuliverFileNameProvider : IFileNameProvider
 {
     public string GetFileNameForSheet(ISheet sheet)
     { 
-        return $"{DateTime.Now:yyyy-MM-dd}_{{CompanyName}}.xlsx";
+        return $"{DateTime.Now:yyyy-MM-dd}_{TemplateBuilderBase.CompanyNameCellFlag}.xlsx";
     }
 }
