@@ -1,4 +1,6 @@
-﻿using System.Text.Json;
+﻿using System.Net;
+using System.Reflection;
+using System.Text.Json;
 
 namespace Kroiko.Domain.TemplateBuilding;
 
@@ -7,11 +9,11 @@ public abstract class TemplateBuilderBase : ITemplateBuilder
     public const string CompanyNameCellFlag = "{CompanyName}";
     public const string MobileNumberCellFlag = "{MobileNumber}";
     public const string TableStartCellFlag = "{TableStart}";
-    public const string DifferentEdgeColorCellFlag = "{DifferentEdgeColor}";
+    public const string MaterialNameCellFlag = "{MaterialName}";
 
-    public abstract Task<IList<ISheet>> BuildTemplateAsync(ContactInfo contactInfo, IEnumerable<KroikoFile> files);
+    public abstract Task<IList<ISheet>> BuildTemplateAsync(IEnumerable<KroikoFile> files);
 
-    protected Cell PopulateStaticInfo(ISheet sheet, ContactInfo contactInfo)
+    protected Cell PopulateStaticInfo(ISheet sheet)
     {
         var tableStartCell = Cell.Empty;
         foreach (var cell in sheet.Cells)
@@ -19,10 +21,10 @@ public abstract class TemplateBuilderBase : ITemplateBuilder
             switch (cell.Value)
             {
                 case CompanyNameCellFlag:
-                    cell.Value = contactInfo.CompanyName;
+                    cell.Value = "GeneratedByKroiko";
                     break;
                 case MobileNumberCellFlag:
-                    cell.Value = contactInfo.MobileNumber;
+                    cell.Value = "GeneratedByKroiko";
                     break;
                 case TableStartCellFlag:
                     tableStartCell = cell;
@@ -43,14 +45,22 @@ public abstract class TemplateBuilderBase : ITemplateBuilder
     
     protected async Task<T> ReadTemplateAsync<T>(string templatePath) where T : ISheet
     {
-        var rawContent = await File.ReadAllTextAsync(templatePath);
-        var templateDefinition = JsonSerializer.Deserialize<T>(rawContent);
-        return templateDefinition;
+        //var rawContent = await File.ReadAllTextAsync(templatePath);
+        //var rawContent = await new HttpClient().GetStringAsync(templatePath);
+        
+        var assembly = Assembly.GetExecutingAssembly();
+        using var stream = assembly.GetManifestResourceStream(templatePath);
+        using var reader = new StreamReader(stream);
+        var jsonString = await reader.ReadToEndAsync();
+        return JsonSerializer.Deserialize<T>(jsonString);
+        
+        // var templateDefinition = JsonSerializer.Deserialize<T>(rawContent);
+        // return templateDefinition;
     }
     
     protected void PopulateDetails(ISheet sheet, Cell tableStartCell, IEnumerable<IKroikoDetail> details, ITableRowProvider tableRowProvider)
     {
-        (int currentRow, int currentColumn) = Cell.GetRowAndColumn(tableStartCell);
+        var (currentRow, currentColumn) = Cell.GetRowAndColumn(tableStartCell);
         foreach (var detail in details)
         {
             var row = tableRowProvider.GetTableRow(detail, currentRow, currentColumn);
