@@ -1,25 +1,27 @@
 ﻿using Kroiko.Domain.ExcelFilesGeneration;
 using Kroiko.Domain.TemplateBuilding;
+using Kroiko.Domain.TemplateBuilding.MegaTrading;
 using System.Text;
 
 namespace Kroiko.Domain.TextFileGeneration;
 
 public interface ITextFileGenerator {
-    List<FileSaveContext> CreateTextBasedFile(ContactInfo contactInfo, IEnumerable<KroikoFile> files);
+    List<FileSaveContext> CreateTextBasedFile(IEnumerable<KroikoFile> files);
 }
-public class MegaTradingFileGenerator : ITextFileGenerator {
-
+public class MegaTradingFileGenerator : ITextFileGenerator 
+{
     // this is a special separator symbol required by the integration destination
     private const string S = "\u256a";
-    public List<FileSaveContext> CreateTextBasedFile(ContactInfo contactInfo, IEnumerable<KroikoFile> files)
+    public List<FileSaveContext> CreateTextBasedFile(IEnumerable<KroikoFile> files)
     {
-        var materials = files.SelectMany(f => f.Details.Cast<MegaTradingDetail>())
+        var kroikoFiles = files as KroikoFile[] ?? files.ToArray();
+        var materials = kroikoFiles.SelectMany(f => f.Details.Cast<MegaTradingDetail>())
             .GroupBy(d => d.Material).ToList();
         var builder = new StringBuilder();
         
         CreateFirstRow(builder);
         
-        // there should always be exactly 6 rows, containing different materials
+        // the integration destination always requires exactly 6 rows, containing different materials
         for (var i = 0; i <= 5; i++)
         {
             if (i >= materials.Count)
@@ -34,14 +36,14 @@ public class MegaTradingFileGenerator : ITextFileGenerator {
         
         CreateColumnSizeRow(builder);
         
-        foreach (var kroikoFile in files)
+        foreach (var kroikoFile in kroikoFiles)
         {
             foreach (MegaTradingDetail detail in kroikoFile.Details)
             {
                 CreateDetailRow(builder, detail);
             }
         }
-        var file = new FileSaveContext($"{contactInfo.CompanyName}.cut_mt", Encoding.UTF8.GetBytes(builder.ToString()));
+        var file = new FileSaveContext("GeneratedByKroiko.cut_mt", Encoding.UTF8.GetBytes(builder.ToString()));
         return [file];
     }
     private static void CreateDetailRow(StringBuilder builder, MegaTradingDetail d)
@@ -52,21 +54,22 @@ public class MegaTradingFileGenerator : ITextFileGenerator {
     }
     private static void CreateColumnSizeRow(StringBuilder builder)
     {
-        // the integration destination uses a GridView to visualize all details
+        // the integration destination uses a WinForms GridView to visualize all details
         // this is the row defining the size of each column
         builder.AppendLine(
         $"50{S}220{S}80{S}80{S}50{S}50{S}90{S}90{S}90{S}90{S}200{S}200{S}");
     }
-    private static void CreateMaterialRow(StringBuilder builder, MegaTradingDetail? detail = null)
-    {
-        // there should always be exactly 6 rows, containing different materials
-        var material = detail == null ? string.Empty : detail.Material;
-        var thickness = detail == null ? 18.0 : detail.Thickness;
-        builder.AppendLine($"{material}{S}{thickness}{S}True{S}True{S}2800{S}2070{S}");
-    }
     private static void CreateFirstRow(StringBuilder builder)
     {
         // holds a boolean value indicating if the material of the edge banding is different from material itself
+        // when parsed by the destination software, this results in a checked box in the UI, removing the need
+        // to define edge banding material for each separate detail
         builder.AppendLine($"{S}{S}True");
+    }
+    private static void CreateMaterialRow(StringBuilder builder, MegaTradingDetail? detail = null)
+    {
+        var material = detail == null ? string.Empty : detail.Material;
+        var thickness = detail?.Thickness ?? 18.0;
+        builder.AppendLine($"{material}{S}{thickness}{S}True{S}True{S}2800{S}2070{S}");
     }
 }
